Crew Document changes

+++ src/main/java/com/intuitbrains/dao/crew/DocumentRepository.java
@@ -1,9 +1,9 @@
 package com.intuitbrains.dao.crew;
 
 
-import com.intuitbrains.model.common.document.category.CrewDocument;
+import com.intuitbrains.model.crew.CrewDocument;
 import org.springframework.data.mongodb.repository.MongoRepository;
 
-public interface DocumentRepository extends MongoRepository<CrewDocument, Integer> {
+public interface DocumentRepository extends MongoRepository<CrewDocument, Long> {
 
 }
